Added manifest to deploy RTR service broker
In the context of
https://app.clubhouse.io/springernature/story/5019/implement-rtr-opt-in
we need to deploy a new service broker that will send logs to port 5515
of logstash.

Sending the logs to that port will enable RTR logs.

By adding new manifest files, we can deploy other SBs with the same
code, but that sends the logs to a different port of logstash.

Pipeline can be found here: https://concourse.halfpipe.io/teams/engineering-enablement/pipelines/ee-app-logging-servicebroker-onPrem